backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pyyaml
pydantic
hypothesis
pytest
httpx
pytest-mock
cryptography
